fixed getclientsfromtopic
