+++ CleanChat.API/Controllers/ClientController.cs
@@ -53,5 +53,25 @@ namespace CleanChat.API.Controllers
             return Ok(ResponseHandler.GetApiResponse(ResponseType.Success,response));
         }
 
+        [HttpPost("topic")]
+        public ActionResult<SubscribeTopicResponse> SubscribeTopic(SubscribeTopicRequest request)
+        {
+            try
+            {
+                var result = _services.SubscribeTopic(request);
+                if (result.Status == null)
+                {
+                    return NotFound(ResponseHandler.GetApiResponse(ResponseType.NotFound, request));
+                } else if (result.Status == false)
+                {
+                    return Ok(ResponseHandler.GetApiResponse(ResponseType.AlreadyExist, request));
+                }
+                return Ok(ResponseHandler.GetApiResponse(ResponseType.Success, result));
+            }
+            catch (Exception e)
+            {
+                return BadRequest(ResponseHandler.GetApiResponse(ResponseType.Failure, e));
+            }
+        }
     }
 }
