+++ CleanChat.API/Program.cs
@@ -36,9 +36,6 @@ builder.Services.AddScoped<ITopicService, TopicService>();
 builder.Services.AddScoped<IClientRepository, ClientRepository>();
 builder.Services.AddScoped<IClientServices, ClientServices>();
 
-builder.Services.AddScoped<IClientTopicRepository, ClientTopicRepository>();
-builder.Services.AddScoped<IClientTopicService, ClientTopicService>();
-
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
